\name{gomms.plot}
\alias{gomms.plot}
\title{
Plot Factor Loadings
}
\description{
plot estimated factor loadings for each sample.
}
\usage{
gomms.plot(X, Y, col.markers = NULL, pch.markers = NULL, ...)
}
\arguments{
  \item{X}{two dimnsional matrix of factor scores.}
  \item{Y}{one or two dimensional matrix of classification.}
  \item{col.markers}{user specified colors for classification.}
  \item{pch.markers}{user specified plot symbols for classification.}
  \item{...}{optional graphical parameters to be passed.}
}