\name{Qqpois}
\alias{Qqpois}
\title{
Probability of a Zero from a Zero State
}
\description{
estimate the probability of a zero from a zero state.
}
\usage{
Qqpois(cdat, eta.hat, mu.hat, dispersion)
}
\arguments{
  \item{cdat}{count Data.}
  \item{eta.hat}{estimated proportion of zeros from a zero state.}
  \item{mu.hat}{estimated mean count.}
  \item{dispersion}{estimated values for dispersion.}
}
\value{
estimated probability of a zero from a zero state.
}